Build generic api response and live data adapter for retrofit calls
The GenericApiResponse will be used to wrap around the response from
retrofit. This is done to allow us to access more information from it
other than just the data from the http response. Like if it had an empty
body, if it was successful or not etc.

The LiveDataCallAdapter and its factory are classes that can always stay
this way and can be copy pasted from this project to another when we
need this functionality. The class itself is responsible for turning the
Call object that the retrofit response normally returns into a LiveData
object. The factory is responsible for creating the object of the class

The ApiService interface has been updated to use those new features and
now returns a LiveData object which contains the result wrapped in the
GenericApiResponse class.