Build the repository class
The repository class will be responsible for getting requests of
getting users or blogs, and will return a LiveData object which contains
MainViewState. This way it can inside there contain information about
the variable that we are interested in, for example when asking for a
user it contains only information about the user and when calling for a
blogpost it will contain only information about that, shown in the part
of the code (user = apiResponse.body or blogPosts = apiResponse.body).