Update the code to make use of the new DataState class
Wrap all times we are simply using MainViewState with our new DataState
class which allows us to know more about the result, like if there was
an error, if there is data on it or if there must be loading or not
(which is currently not used yet).

Update the repository class to handle the data with this in mind, by
wrapping the responses appropriately.

Update the ViewModel to use the DataState as well.

Finally update the MainFragment to properly subscribe to the changes
made to the DataState object, and create code for handling all different
cases like if there is data there, handle it as we did before, but now
also create methods to put code in the future to handle if there is an
error message in it, or update whether there is something loading or
not.