Add DataStateListener interface
This class will be used to transfer information from fragments to
activities. Particularly, information about the DataState and when it
changes. This results in the activity knowing about if there is a
message inside the data, because an error has happened for example
which can be used for showing a Toast. And also, if the state is at a
state of loading, therefore update the UI accordingly.