Add Recycler adapter using DiffUtils for recycler view
Add initialization of the recycler view inside the main fragment. Also
add update of the list when the viewState is updated, by submitting
the new blog-posts when a change is observed.
Update the BlogPost model class so that it supports the .equals method.
Use the ListAdapter template class to implement the diff-utils
implementation for the recycler view.
TopSpacingItemDecoration is a simple item decoration extending class
that will simply put a padding on top of each item.